backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "character-battle-backend"
version = "0.1.0"
description = "FastAPI backend for the character battle service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "redis>=5.0.0",
  "pydantic-settings>=2.4.0",
  "openai>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
